adds webpack to split leetcode.js
