+++ scripts/leetcode/util.js
@@ -1,0 +1,95 @@
+/* Enum for languages supported by LeetCode. */
+const languages = {
+  C: '.c',
+  'C++': '.cpp',
+  'C#': '.cs',
+  Dart: '.dart',
+  Elixir: '.ex',
+  Erlang: '.erl',
+  Go: '.go',
+  Java: '.java',
+  JavaScript: '.js',
+  Javascript: '.js',
+  Kotlin: '.kt',
+  MySQL: '.sql',
+  'MS SQL Server': '.sql',
+  Oracle: '.sql',
+  Pandas: '.py',
+  PHP: '.php',
+  Python: '.py',
+  Python3: '.py',
+  Racket: '.rkt',
+  Ruby: '.rb',
+  Rust: '.rs',
+  Scala: '.scala',
+  Swift: '.swift',
+  TypeScript: '.ts',
+};
+
+const DIFFICULTY = Object.freeze({
+    EASY: 'Easy',
+    MEDIUM: 'Medium',
+    HARD: 'Hard',
+    UNKNOWN: 'Unknown'
+})
+
+function getBrowser() {
+    if (typeof chrome !== "undefined" && typeof chrome.runtime !== "undefined") {
+      return chrome;
+    } else if (typeof browser !== "undefined" && typeof browser.runtime !== "undefined") {
+      return browser;
+    } else {
+      console.log("browser is not supported");
+      return false;
+    }
+  }
+
+function getDifficulty(difficulty) {
+    difficulty = difficulty.toUpperCase().trim()
+    return DIFFICULTY[difficulty] ?? DIFFICULTY.UNKNOWN
+}
+
+/* Checks if an elem/array exists and has length */
+function checkElem(elem) {
+  return elem && elem.length > 0;
+}
+
+function convertToSlug(string) {
+  const a = 'àáâäæãåāăąçćčđďèéêëēėęěğǵḧîïíīįìłḿñńǹňôöòóœøōõőṕŕřßśšşșťțûüùúūǘůűųẃẍÿýžźż·/_,:;';
+  const b = 'aaaaaaaaaacccddeeeeeeeegghiiiiiilmnnnnoooooooooprrsssssttuuuuuuuuuwxyyzzz------';
+  const p = new RegExp(a.split('').join('|'), 'g');
+
+  return string
+    .toString()
+    .toLowerCase()
+    .replace(/\s+/g, '-') // Replace spaces with -
+    .replace(p, c => b.charAt(a.indexOf(c))) // Replace special characters
+    .replace(/&/g, '-and-') // Replace & with 'and'
+    .replace(/[^\w\-]+/g, '') // Remove all non-word characters
+    .replace(/\-\-+/g, '-') // Replace multiple - with single -
+    .replace(/^-+/, '') // Trim - from start of text
+    .replace(/-+$/, ''); // Trim - from end of text
+}
+
+function addLeadingZeros(title) {
+  const maxTitlePrefixLength = 4;
+  var len = title.split('-')[0].length;
+  if (len < maxTitlePrefixLength) {
+    return '0'.repeat(4 - len) + title;
+  }
+  return title;
+}
+
+function formatStats(time, timePercentile, space, spacePercentile) {
+  return `Time: ${time} (${timePercentile}%), Space: ${space} (${spacePercentile}%) - LeetHub`;
+}
+
+export {
+  checkElem,
+  convertToSlug,
+  languages,
+  addLeadingZeros,
+  formatStats,
+  getDifficulty,
+  DIFFICULTY
+};
